+++ central/billing/api/dashboard/__init__.py
@@ -20,11 +20,7 @@ from central.billing.api.dashboard.account import (
 	get_collection_status,
 	get_team_overview,
 	get_trust_tier,
-	list_notifications,
 	list_switchable_teams,
-	mark_all_notifications_read,
-	mark_notification_read,
-	notification_badge,
 	save_billing_profile,
 	save_billing_settings,
 	set_collection_mode,
@@ -88,14 +84,10 @@ __all__ = [
 	"get_trust_tier",
 	"initiate_card_setup",
 	"list_invoices",
-	"list_notifications",
 	"list_payment_attempts",
 	"list_payment_methods",
 	"list_subscriptions",
 	"list_switchable_teams",
-	"mark_all_notifications_read",
-	"mark_notification_read",
-	"notification_badge",
 	"pause_subscription",
 	"pay_invoice",
 	"pay_invoice_checkout",
