+++ central/billing/tests/test_dashboard.py
@@ -807,8 +807,6 @@ class TestWriteEndpointsRejectGet(IntegrationTestCase):
 			methods.reorder_payment_methods,
 			account.save_billing_profile,
 			account.save_billing_settings,
-			account.mark_notification_read,
-			account.mark_all_notifications_read,
 		]
 		allowed = frappe.allowed_http_methods_for_whitelisted_func
 		for fn in write_fns:
